# doc/api.yaml
openapi: 3.0.0
paths:
...
  /track:
    post:
      description: This endpoint creates a new track with a name, year, and an
        optional artist ID.
      operationId: TracksController_addTrack
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "201":
          description: The track has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request. The input data is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - year must not be greater than 2025
                    - year must not be less than 1900
                    - year must be an integer number
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Artist or album not exist. The artist or album with the specified
            ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: &a7
        - access-token: []
      summary: Create a new track
      tags: &a8
        - Tracks
    get:
      description: This endpoint retrieves a list of all tracks.
      operationId: TracksController_getTracks
      parameters: []
      responses:
        "200":
          description: A list of tracks has been successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
      security: *a7
      summary: Get all tracks
      tags: *a8
  /track/{id}:
    get:
      description: This endpoint retrieves an track by its ID.
      operationId: TracksController_getTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "200":
          description: The track has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
      security: *a7
      summary: Get an track by ID
      tags: *a8
    put:
      description: This endpoint updates an existing track by its ID with new data.
      operationId: TracksController_updateTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: The The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
        "422":
          description: Artist or album not exist. The artist or album with the specified
            ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: *a7
      summary: Update an track by ID
      tags: *a8
    delete:
      description: This endpoint deletes an track by its ID.
      operationId: TracksController_deleteTrack
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "204":
          description: The track has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found
                  error: Not Found
      security: *a7
      summary: Delete an track by ID
      tags: *a8
...
  /favs/track/{id}:
    post:
      description: This endpoint adds a track to the favorites.
      operationId: FavoritesController_addTrackToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "201":
          description: The track has been successfully added to favorites.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddedFavorite"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Track not exist. The track with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 does not exist
                  error: Not Found
      security: *a9
      summary: Add a track to favorites
      tags: *a10
    delete:
      description: This endpoint removes a track from the favorites.
      operationId: FavoritesController_removeTrackFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the track (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174003
            type: string
      responses:
        "204":
          description: The track has been successfully removed from favorites.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Track not found in favorites. The track with the specified ID does
            not exist in favorites.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Track with ID 123e4567-e89b-12d3-a456-426614174003 not found in
                    favorites
                  error: Not Found
      security: *a9
      summary: Remove a track from favorites
      tags: *a10
...
info:
  title: Home Library Service
  description: API for managing music albums, artists, and tracks
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      description: Enter JWT access token
      in: header
  schemas:
...
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the track
          example: Come Together
        duration:
          type: number
          description: The duration of the track in seconds
          example: 201
        artistId:
          type: string
          description: The artist ID associated with the track
          nullable: true
          format: uuid
        albumId:
          type: string
          description: The album ID associated with the track
          nullable: true
          format: uuid
      required:
        - name
        - duration
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174003
        name:
          type: string
          example: Come Together
        duration:
          type: number
          example: 201
        artistId:
          type: string
          description: The artist ID associated with the track
          nullable: true
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        albumId:
          type: string
          description: The album ID associated with the track
          nullable: true
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the track
          example: Come Together
        duration:
          type: number
          description: The duration of the track in seconds
          example: 201
        artistId:
          type: string
          description: The artist ID associated with the track
          nullable: true
          format: uuid
        albumId:
          type: string
          description: The album ID associated with the track
          nullable: true
          format: uuid
...
    AddedFavorite:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
          description: The unique identifier of the added favorite entity
        type:
          type: string
          enum:
            - artist
            - album
            - track
          example: artist
          description: The type of the added favorite entity
...